requirements.txt
langchain==0.2.8
langchain_community==0.2.7
langchain_core==0.2.20
langchain_openai==0.1.16
langchain_text_splitters==0.2.2
pandas==2.2.2
pdfplumber==0.11.2
python-dotenv==1.0.1
streamlit==1.35.0
